Modularize and fix toggle.
