+++ package.json
@@ -30,7 +30,7 @@
     "stylelint": "^13.3.3",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.17.2",
+    "stylelint-scss": "^3.19.0",
     "webpack": "^5.21.2",
     "webpack-cli": "^4.5.0"
   },
